fix: 🐛 外場送出餐點紀錄, socket 回給廚房端資料修正與 table-list 相同
